+++ types/react.d.ts
@@ -83,7 +83,7 @@ declare module 'react' {
       'm-command-palette': MicaElement<{ filter?: 'manual' }>
       'm-command-list': MicaElement<{ dividers?: boolean }>
       'm-command-group': MicaElement
-      'm-combobox': MicaElement
+      'm-combobox': MicaElement & { filter?: 'manual' }
     }
   }
 }